New Feature to allow only mapvotes that correspond with the mappool.
cmd_mappool can easily spam if mappool.txt is huge - maybe use another format etc? Also im not too happy with the file handling in __init__ - we just try to read from basepath first, if it fails it fails. we then read from homepath, eventually overwriting the mappool.txt we fetched from basepath.

Other than that, I tried to make the functions as robust as possible to account for votes like /callvote MaP CamPGrounDs and stuff. checkMapvoteAllowed() checks if the map is in the mappool and if the factory(if any) can be applied. 

i named the cvar "qlx_forceMappool"